// clients/js-kit/src/generated/types/metadataDelegateRole.ts
/**
 * This code was AUTOGENERATED using the Codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun Codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import {
  combineCodec,
  getEnumDecoder,
  getEnumEncoder,
  type FixedSizeCodec,
  type FixedSizeDecoder,
  type FixedSizeEncoder,
} from '@solana/kit';

export enum MetadataDelegateRole {
  AuthorityItem,
  Collection,
  Use,
  Data,
  ProgrammableConfig,
  DataItem,
  CollectionItem,
  ProgrammableConfigItem,
}

export type MetadataDelegateRoleArgs = MetadataDelegateRole;

export function getMetadataDelegateRoleEncoder(): FixedSizeEncoder<MetadataDelegateRoleArgs> {
  return getEnumEncoder(MetadataDelegateRole);
}

export function getMetadataDelegateRoleDecoder(): FixedSizeDecoder<MetadataDelegateRole> {
  return getEnumDecoder(MetadataDelegateRole);
}

export function getMetadataDelegateRoleCodec(): FixedSizeCodec<
  MetadataDelegateRoleArgs,
  MetadataDelegateRole
> {
  return combineCodec(
    getMetadataDelegateRoleEncoder(),
    getMetadataDelegateRoleDecoder()
  );
}